dri2: Fix double-sync of pageflips on Linux 3.13+ - Part I
Linux 3.13 and later sync kms pageflips to vblank in the kms
driver, so we must not emit a sync to vblank pushbuf in the ddx
on such kernels, or maximum framerate will be cut into half!

A sync-to-vblank-pushbuf is emitted for copyswaps as in the past,
also for older kernels which don't support async_pageflip's and
don't sync by themselves.

This adds the implementation, but not the detection logic for
async_pageflip support in the kernel.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>