Fix event handling on multi-x-screen configs.
Only register the wakeup handler and event socket
once per card fd and server generation, as the fd
and device file is shared between all x-screens for
a given card during a given server generation.

Without this fix, vblank and kms-pageflip completion
event processing don't work properly, as the server
doesn't kick the wakeup handler for gpu events, and
therefore the desktop will freeze, unless the user
manually kicks the wakeup handler by moving the mouse
or hitting the keyboard.

Add proper reference counting and checks to make it so.
This fix is derived from a similar and proven fix in
the ati ddx for the same problem.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>