xv/dri2/dri3: Switch to optimized crtc selection for drawables.
Make crtc selection consistent with ati and intel ddx.

Pick the crtc to use for vblank events, swap scheduling and
kms-pageflip completion events as the one with maximum pixel area
intersection between its viewport and the drawable. If multiple
crtc's viewports display the same amount of area of a drawable,
select the crtc whose output is the designated "primary output"
for the x-screen. The latter allows, e.g., in a clone display
setup, to select the important output on which tearing should be
minimized. The former makes sure to minimize or avoid tearing on
the display which shows the largest part of the drawable.

This method is transplanted from the ati ddx.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>