dri2: Set default SwapLimit to 1 for double-buffering.
This used to be 2 for pseudo-triplebuffering on XOrg 1.12+.

However, it caused problems like unthrottled swapping
at up to twice the video refresh rate for redirected
windows under desktop composition, which can't be easily
fixed under dri2.

Default to double-buffering. Users can override this, but
dri3 provides better solutions for this.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>